+++ Assets/Art/Character/Prefab/Reptile.prefab
@@ -693,6 +693,8 @@ GameObject:
   - component: {fileID: 136745343930619558}
   - component: {fileID: 114142692803894372}
   - component: {fileID: 114578420604300796}
+  - component: {fileID: 82354351781239318}
+  - component: {fileID: 114538764986964210}
   m_Layer: 0
   m_Name: Reptile
   m_TagString: Player
@@ -1749,6 +1751,101 @@ AudioListener:
   m_PrefabInternal: {fileID: 100100000}
   m_GameObject: {fileID: 1900615736734028}
   m_Enabled: 1
+--- !u!82 &82354351781239318
+AudioSource:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 100090}
+  m_Enabled: 1
+  serializedVersion: 4
+  OutputAudioMixerGroup: {fileID: 0}
+  m_audioClip: {fileID: 0}
+  m_PlayOnAwake: 1
+  m_Volume: 1
+  m_Pitch: 1
+  Loop: 0
+  Mute: 0
+  Spatialize: 0
+  SpatializePostEffects: 0
+  Priority: 128
+  DopplerLevel: 1
+  MinDistance: 1
+  MaxDistance: 500
+  Pan2D: 0
+  rolloffMode: 0
+  BypassEffects: 0
+  BypassListenerEffects: 0
+  BypassReverbZones: 0
+  rolloffCustomCurve:
+    serializedVersion: 2
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 1
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.33333334
+      outWeight: 0.33333334
+    - serializedVersion: 3
+      time: 1
+      value: 0
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.33333334
+      outWeight: 0.33333334
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
+  panLevelCustomCurve:
+    serializedVersion: 2
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 0
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.33333334
+      outWeight: 0.33333334
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
+  spreadCustomCurve:
+    serializedVersion: 2
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 0
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.33333334
+      outWeight: 0.33333334
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
+  reverbZoneMixCustomCurve:
+    serializedVersion: 2
+    m_Curve:
+    - serializedVersion: 3
+      time: 0
+      value: 1
+      inSlope: 0
+      outSlope: 0
+      tangentMode: 0
+      weightedMode: 0
+      inWeight: 0.33333334
+      outWeight: 0.33333334
+    m_PreInfinity: 2
+    m_PostInfinity: 2
+    m_RotationOrder: 4
 --- !u!114 &114142692803894372
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1760,11 +1857,11 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 8e78e05358b7e68499f62fcfaf92e19a, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  rightHandCollider: {fileID: 65900574269286190}
-  leftHandCollider: {fileID: 65469049637776126}
   cameraPivot: {fileID: 1662844323145896}
   walkSpeed: 15
   runSpeed: 30
+  woosh: {fileID: 0}
+  hit: {fileID: 0}
 --- !u!114 &114332298484477060
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -1774,6 +1871,17 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 5006cf63c28b8b843900eb80672fcbbb, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &114538764986964210
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 100090}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 922135b9453f7d640b834092c4b1bb6c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!114 &114578420604300796
@@ -1790,6 +1898,7 @@ MonoBehaviour:
   maxHealth: 100
   maxStamina: 100
   baseDamage: 10
+  m_Renderer: {fileID: 0}
 --- !u!124 &124828548360769644
 Behaviour:
   m_ObjectHideFlags: 1
