+++ Assets/Art/Low Poly Animated People/- Prefabs/grandma.prefab
@@ -160,6 +160,7 @@ GameObject:
   - component: {fileID: 114855439647082994}
   - component: {fileID: 95448043312301298}
   - component: {fileID: 82445394445386236}
+  - component: {fileID: 114669839475564572}
   m_Layer: 0
   m_Name: grandma
   m_TagString: Untagged
@@ -903,6 +904,17 @@ MonoBehaviour:
   scoreAwarded: 5
   scream: {fileID: 8300000, guid: 715b6b61d574bc6469aba6d91097e106, type: 3}
   splat: {fileID: 8300000, guid: a4676e0119a23ce458ff7530a8a4f7cf, type: 3}
+--- !u!114 &114669839475564572
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1274296657605726}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: de88fca8291a1c445950ce95f891d040, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &114855439647082994
 MonoBehaviour:
   m_ObjectHideFlags: 1
