+++ Assets/Art/Low Poly Animated People/- Prefabs/man.prefab
@@ -205,6 +205,7 @@ GameObject:
   - component: {fileID: 114599641657491824}
   - component: {fileID: 95863098599339194}
   - component: {fileID: 82495145987908206}
+  - component: {fileID: 114636587219547110}
   m_Layer: 0
   m_Name: man
   m_TagString: Human
@@ -895,6 +896,17 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   distance: 5
   throwingForce: 100
+--- !u!114 &114636587219547110
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1503856157617204}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: de88fca8291a1c445950ce95f891d040, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &114754278264576922
 MonoBehaviour:
   m_ObjectHideFlags: 1
