+++ Assets/Art/Misc Prefabs/WaterSpray.prefab
@@ -23,6 +23,7 @@ GameObject:
   - component: {fileID: 199335414687463910}
   - component: {fileID: 114277414402833948}
   - component: {fileID: 82785373871335038}
+  - component: {fileID: 114493599037577884}
   m_Layer: 0
   m_Name: WaterSpray
   m_TagString: Untagged
@@ -150,6 +151,17 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   timeToLive: 5
+--- !u!114 &114493599037577884
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1363128862032008}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: de88fca8291a1c445950ce95f891d040, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!198 &198960461315985660
 ParticleSystem:
   m_ObjectHideFlags: 1
