--- Assets/Scripts/SoundEffect.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 922135b9453f7d640b834092c4b1bb6c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
