+++ Assets/Scripts/UI/Settings/Audio/SoundEffectSource.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: de88fca8291a1c445950ce95f891d040
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
