--- Assets/Scripts/UI/Settings/Audio/SoundEffects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9adbef7d29051fb46bf28316731f8654
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
